Berg v1.0.0 Berg

A Minimum Heap for Integers

Types

element()

element() :: integer

An Element (just an Integer)

t()

t

A Minimum Heap

Functions

extract(x)

extract(Berg.t) :: {Berg.t, element}

A heap alongside an element where that element would be the smallest if it were in the heap

heapify(x)

heapify([integer]) :: Berg.t

A heap with the same elements as the list in question

insert(x, y)

insert(Berg.t, element) :: Berg.t

A heap whose smallest element is the minimum of the element in question and the smallest element of the heap in question

listify(x)

listify(Berg.t) :: [integer]

A list in ascending order (w/ all the same elements as the heap)

root(x)

root(Berg.t) :: element

The smallest element of the heap

trunk(x)

The heap in question w/o the smallest element

unary(x)

unary(element) :: Berg.t

A Heap w/ just one element

zero()

zero() :: Berg.t

A zero heap

zero?(x)

zero?(Berg.t) :: boolean

Predicate: is this the zero heap?