Berg v2.1.0 Berg.Heap

Find primitives in this module

Types

t()

t

A Minimum Heap (of Integers)

Functions

extract(x)

extract(Berg.Heap.t) :: {Berg.Heap.t, Berg.Heap.Element.t}

A heap alongside an element where that element would be the smallest if it were in the heap

Example

iex> alias Berg
iex> alias Berg.Heap
iex> l = [42, 16, 8, 4, 15, 23]
iex> h = Berg.heapify(l)
iex> {t, e} = Heap.extract(h)
iex> Berg.listify(t)
[8, 15, 16, 23, 42]
iex> e
4

insert(x, y)

insert(Berg.Heap.t, Berg.Heap.Element.t) :: Berg.Heap.t

A heap whose smallest element is the minimum of the element in question and the smallest element of the heap in question

Example

iex> alias Berg
iex> alias Berg.Heap
iex> l = [16, 8, 15, 23]
iex> h = Berg.heapify(l)
iex> i = h |> Heap.insert(4) |> Heap.insert(42)
iex> Heap.root(i)
4

root(x)

root(Berg.Heap.t) :: Berg.Heap.Element.t

The smallest element of the heap

Example

iex> alias Berg
iex> alias Berg.Heap
iex> l = [42, 16, 8, 4, 15, 23]
iex> h = Berg.heapify(l)
iex> Heap.root(h)
4

trunk(x)

trunk(Berg.Heap.t) :: Berg.Heap.t

The heap in question w/o the smallest element

Example

iex> alias Berg
iex> alias Berg.Heap
iex> l = [42, 16, 8, 4, 15, 23]
iex> h = Berg.heapify(l)
iex> t = Heap.trunk(h)
iex> Berg.listify(t)
[8, 15, 16, 23, 42]

unary(x)

unary(Berg.Heap.Element.t) :: Berg.Heap.t

A Heap w/ just one element

Example

iex> alias Berg.Heap
iex> Heap.zero?(Heap.unary(4))
false

zero()

zero() :: Berg.Heap.t

A zero heap

zero?(x)

zero?(Berg.Heap.t) :: boolean

Predicate: is this the zero heap?

Example

iex> alias Berg.Heap
iex> Heap.zero?(Heap.zero())
true